fix: brightness is now using non linear equation

+++ src/tools/Greyscale.h
@@ -11,13 +11,13 @@ public:
   static void convertToGreyscale(std::unique_ptr<Image>& image);
 
   // Funkcje modyfikujące obraz przy użyciu LUT
-  static void adjustBrightness(std::unique_ptr<Image>& image, int value);
+  static void adjustBrightness(std::unique_ptr<Image>& image, float value);
   static void adjustContrast(std::unique_ptr<Image>& image, float factor);
   static void adjustGamma(std::unique_ptr<Image>& image, float gamma);
 
 private:
   // Funkcje tworzące tablice LUT
-  static std::array<int, 256> createBrightnessLUT(int value);
+  static std::array<int, 256> createBrightnessLUT(float value);
   static std::array<int, 256> createContrastLUT(float factor);
   static std::array<int, 256> createGammaLUT(float gamma);
